requirements.txt
Flask
python-docx
reportlab
arabic-reshaper
python-bidi
gunicorn
